conditional for aws

+++ src/app.js
@@ -91,16 +91,20 @@ app.use('/', router);
 console.log(process.env.NODE_ENV)
 
 if (process.env.NODE_ENV === 'production') {
-	fs.unlink(process.env.PORT, () => {
-		console.log('cleared old socket');
+    if (process.env.TARGET === 'aws') {
+        app.listen(80, () => { console.log('Serving on 80') });
+    } else {
+        fs.unlink(process.env.PORT, () => {
+    		console.log('cleared old socket');
 
-        app.listen(process.env.PORT, () => {
-        	console.log('listening on unix socket');
+            app.listen(process.env.PORT, () => {
+            	console.log('listening on unix socket');
 
-        	fs.chmodSync(process.env.PORT, '777');
-        	console.log('set permissions of socket to 777');
-        });
-	});
+            	fs.chmodSync(process.env.PORT, '777');
+            	console.log('set permissions of socket to 777');
+            });
+    	});
+    }
 } else {
 	app.listen(8000, () => {
 		console.log('listening on 8000');
